Add challenge expiresAt response field (#34)
* Add expiresAt field to ChallengeResponse and update type to Long

* Bump version to 2.8.0

+++ src/main/java/com/authsignal/model/GetChallengeResponse.java
@@ -2,7 +2,7 @@ package com.authsignal.model;
 
 public class GetChallengeResponse extends ApiModel {
     public String challengeId;
-    public Integer expiresAt;
+    public Long expiresAt;
     public VerificationMethodType verificationMethod;
     public SmsChannel smsChannel;
     public String phoneNumber;
